Rename to `toString`
